Added format instructions and metadata handling.
Extended the readme to include instructions for using formatters to get
behavior that duplicates older version of `winston-graylog2`. Downside
is that it requires using logform@^2.1.0, which hasn't been added to
base winston yet.

Also included a little code for including the metadata in the log
output.

Lastly, I noticed that the winston documentation suggests wrapping the
call to the actual transport client in a setImmediate, so added that.